+++ lib/src/widget/khqr_card_widget.dart
@@ -106,7 +106,6 @@ class _KhqrCardWidgetState extends State<KhqrCardWidget> {
   @override
   void initState() {
     super.initState();
-
     WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
       _updateDuration();
     });
@@ -180,6 +179,7 @@ class _KhqrCardWidgetState extends State<KhqrCardWidget> {
                   child: SvgPicture.asset(
                     'assets/svg/khqr_logo.svg',
                     package: 'khqr_sdk',
+                    colorFilter: ColorFilter.mode(_pearlWhite, BlendMode.srcIn),
                   ),
                 ),
                 Expanded(
@@ -187,8 +187,9 @@ class _KhqrCardWidgetState extends State<KhqrCardWidget> {
                     width: widget.width,
                     color: _bakongBraveryRed,
                     child: ClipPath(
-                      clipper:
-                          _KhqrCardHeaderClipper(aspectRatio: _aspectRatio),
+                      clipper: _KhqrCardHeaderClipper(
+                        aspectRatio: _aspectRatio,
+                      ),
                       child: Container(
                         color: qrBackgroundColor,
                         child: Column(
@@ -239,10 +240,17 @@ class _KhqrCardWidgetState extends State<KhqrCardWidget> {
                                             color: qrTextColor,
                                           ),
                                         )
-                                      : Text(''),
-                                  if (widget.amount > 0 ||
-                                      widget.showEmptyAmount)
-                                    SizedBox(width: _height * 0.02),
+                                      : Text(
+                                          '',
+                                          style: TextStyle(
+                                            fontSize: _amountFontSize,
+                                          ),
+                                        ),
+                                  Visibility(
+                                    visible: widget.amount > 0 ||
+                                        widget.showEmptyAmount,
+                                    child: SizedBox(width: _height * 0.02),
+                                  ),
                                   Text(
                                     widget.currency.name.toUpperCase(),
                                     style: TextStyle(